# services/facade-api/openapi.yaml
openapi: 3.1.0
info:
  title: Seftec Unified Finance API
  description: |
    Unified API for wallet management, transfers, payments, cards, and KYC verification.

    ## Authentication
    All requests require an API key passed in the `X-API-Key` header.

    ## Idempotency
    For POST requests, include an `Idempotency-Key` header to prevent duplicate operations.
    The key must be unique per request and will be valid for 24 hours.

    ## Webhooks
    Register webhook endpoints to receive real-time notifications for:
    - `transfer.completed`, `transfer.failed`
    - `payment.successful`, `payment.failed`
    - `wallet.credited`, `wallet.debited`
    - `kyc.verified`, `kyc.failed`

    ## Rate Limits
    Rate limits depend on your tier:
    - Starter: 60 req/min, 500 req/hour
    - Growth: 120 req/min, 2000 req/hour
    - Business: 300 req/min, 10000 req/hour
    - Enterprise: Custom

  version: "1.0.0"
  contact:
    name: Seftec API Support
    email: api-support@seftechub.com
    url: https://developers.seftechub.com
  license:
    name: Proprietary
    url: https://seftechub.com/terms
servers:
  - url: https://api.seftechub.com/v1
    description: Production
  - url: https://sandbox.seftechub.com/v1
    description: Sandbox (test mode)
security:
  - ApiKeyAuth: []
...
paths:
...
  /wallets:
    post:
      operationId: createWallet
      tags: [Wallets]
      summary: Create a wallet
      description: |
        Creates a new wallet for a customer. Each customer can have multiple wallets
        in different currencies.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customer_id, currency]
              properties:
                customer_id:
                  type: string
                  format: uuid
                  description: Your customer's unique identifier
                currency:
                  $ref: "#/components/schemas/Currency"
                type:
                  type: string
                  enum: [personal, business, savings]
                  default: personal
                metadata:
                  type: object
                  additionalProperties: true
            example:
              customer_id: "cust_abc123"
              currency: "NGN"
              type: "personal"
              metadata:
                source: "mobile_app"
      responses:
        "201":
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"

    get:
      operationId: listWallets
      tags: [Wallets]
      summary: List wallets
      parameters:
        - name: customer_id
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: List of wallets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletListResponse"
  /wallets/{wallet_id}:
    get:
      operationId: getWallet
      tags: [Wallets]
      summary: Get wallet details
      parameters:
        - name: wallet_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Wallet details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /wallets/{wallet_id}/balance:
    get:
      operationId: getWalletBalance
      tags: [Wallets]
      summary: Get wallet balance
      parameters:
        - name: wallet_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Wallet balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceResponse"
...
  /webhooks:
    post:
      operationId: createWebhook
      tags: [Webhooks]
      summary: Register webhook endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                  description: Events to receive (empty = all)
                description:
                  type: string
            example:
              url: "https://yourapp.com/webhooks/seftec"
              events: ["transfer.completed", "transfer.failed", "payment.successful"]
      responses:
        "201":
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"

    get:
      operationId: listWebhooks
      tags: [Webhooks]
      summary: List webhook endpoints
      responses:
        "200":
          description: List of webhooks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookListResponse"
  /webhooks/{webhook_id}:
    delete:
      operationId: deleteWebhook
      tags: [Webhooks]
      summary: Delete webhook endpoint
      parameters:
        - name: webhook_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Webhook deleted
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Your project API key
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Unique key to prevent duplicate operations (recommended for POST requests)
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Pagination cursor from previous response
  schemas:
    Currency:
      type: string
      enum: [NGN, USD, GBP, EUR, GHS, KES, ZAR]
...
    Wallet:
      type: object
      properties:
        id:
          type: string
        customer_id:
          type: string
        type:
          type: string
        currency:
          $ref: "#/components/schemas/Currency"
        status:
          type: string
          enum: [active, frozen, suspended, closed]
        account_number:
          type: string
        account_name:
          type: string
        bank_name:
          type: string
        created_at:
          type: string
          format: date-time
    Balance:
      type: object
      properties:
        available:
          type: number
          description: Available for spending/transfer
        ledger:
          type: number
          description: Total including pending
        reserved:
          type: number
          description: Reserved for pending operations
        currency:
          $ref: "#/components/schemas/Currency"
        last_updated:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
        meta:
          type: object
          properties:
            request_id:
              type: string

    # Response wrappers
    WalletResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Wallet"
        meta:
          type: object
          properties:
            request_id:
              type: string
    WalletListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Wallet"
        meta:
          type: object
          properties:
            cursor:
              type: string
            has_more:
              type: boolean
    BalanceResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Balance"
...
    WebhookResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: string
            url:
              type: string
            events:
              type: array
              items:
                type: string
            secret:
              type: string
              description: Use this to verify webhook signatures
            status:
              type: string
    WebhookListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              url:
                type: string
              events:
                type: array
                items:
                  type: string
              status:
                type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: "INVALID_REQUEST"
              message: "Invalid request body"
              details:
                field: "amount"
                issue: "must be positive"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: "NOT_FOUND"
              message: "Wallet not found"
    Conflict:
      description: Conflict (e.g., duplicate idempotency key)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"